address PR comments, remove old files and adjust rebase
