+++ src/components/forms/sponsor-general-form/add-tier-popup.js
@@ -115,6 +115,9 @@ const AddTierPopup = ({ sponsor, open, onClose, onSubmit, summitId }) => {
                   summitId={summitId}
                   isMulti
                   placeholder={T.translate("edit_sponsor.placeholders.select")}
+                  hiddenOptions={sponsor.sponsorships_collection.sponsorships.map(
+                    (e) => e.type.id
+                  )}
                 />
               </Grid2>
             </Grid2>
