+++ src/components/forms/sponsor-general-form/sponsor-header.js
@@ -16,7 +16,7 @@ import T from "i18n-react/dist/i18n-react";
 import { Box, Divider, Grid2, Typography } from "@mui/material";
 
 const SponsorHeader = ({ sponsor }) => (
-  <Box className="container" sx={{ px: 2, py: 0 }}>
+  <Box sx={{ px: 2, py: 0, backgroundColor: "#FFF" }}>
     <Grid2 container size={12} sx={{ height: "68px", alignItems: "center" }}>
       <Grid2 size={12}>
         <Typography
